refactor: create separate selected items component
